chore: Protocol 및 구현체 네이밍 변경
및 DTO 네이밍 변경
