+++ BoxOffice/Data/Repository/MovieProtocol.swift
@@ -1,0 +1,4 @@
+protocol MovieProtocol {
+  func getDailyBoxOffice(dateString: String) async -> Result<SearchDailyBoxOfficeDTO, MovieRepositoryError>
+  func getMovieDetail(movieCode: String) async -> Result<SearchMovieInfoDTO, MovieRepositoryError>
+}
